RESQUE. Use ":resque" in production, ":inline" in development
Modified the configuration so that the ":resque" queue adapter is used
in the "production" environment, while the ":inline" queue adapter is
used for "development".

The "db:reset" Rake task, used to setup the application, performs
ActiveJob queuing (via the "after_save" callback on the "Vocabulary"
model). Using the ":inline" adapter by default removes the need to have
a running Redis server to setup and test the application. This
simplifies both development setup and the Jenkins CI configuration.

Modified existing instructions in the README to indicate how to modify
the development environment to use the ":resque" adapter when needed.